Simplify OpenSSH mainline build (#2158)
### Description of changes: 
Simplify the build script. No more flags needed for upstream mainline
build. Some flags reduced for v8.9 openssh build.

This is as a result of upstreaming AWS-LC into OpenSSH.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.