Improve tool-openssl compatability for x509 and verify subcommands (#2196)
### Description of changes: 
* OpenSSL supports passing lower-case `pem` or `der` when using the
`-inform` argument to the `x509` sub-command.
* Add support for the `-untrusted` flag to the `verify` sub-command
which allows for passing in a set of one or more chain certificates
which may be used in the certificate chain building and validation.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.