Fix C++98 compatibility in our header files (#2193)
We encountered this when trying to compile Xtrabackup on gcc versions
lower than 10. Apparently scoped enums are a C++11 feature, which the
compiler can't understand without additional linker flags.

We also encountered this in 42a8dba, which prompted us to move it to
`ssl.h`. Upon closer examination, I don't think the C++11 usage is a
necessity especially with the pure C fallback that was included.
Predeclaring enums reduces compilation dependencies and build times,
but our build seems to work fine without it. Removing this would let us
encounter less build issues with projects that aren't expecting C++ in our
header files.

Asides from the issue with scoped enums mentioned above, we decided
to check for C++98 compatibility in CI and fix any potential issues with
`-Wpedantic` while we're at it. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.