Update pairwise consistency test failures to support gracefully continiung (#2201)
### Description of changes: 
Currently AWS_LC_FIPS_failure always aborts the process and it is
impossible to continue after calling that function. In a future change
this will be optional and AWS-LC will not abort and the overall API call
should return a failure. This change updates the current pairwise
consistency test spots to theoretically return a failure.

### Call-outs:
This PR does not change the behavior, if there is a PWCT failure the
process still aborts.

### Testing:
With the current AWS_LC_FIPS_failure implementation it is impossible to
test this change.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.