Remove access() call from Snapsafe detection (#2197)
### Description of changes: 
This change uses stat() instead of access() to check for the existence
of the Sysgenid device. This is more accurate to ascertain whether the
device exists on a machine. access() could fail for various other
reasons such as missing permissions.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.