Modify SSL to inherit ciphersuites from SSL_CTX at initialization (#2198)
### Description of changes: 
This is a follow up to
https://github.com/aws/aws-lc/commit/154f9986bc18f8f0e6126157a3abc84230456f77.
These changes initialize the SSL object with configured ciphersuites
from the parent SSL_CTX object at initialization. Now both SSL and
SSL_CTX objects will have their own copy of ciphersuites.

We inherited the behavior of defaulting to SSL_CTX ciphersuites when
none are set on SSL from BoringSSL. OpenSSL does things differently and
copies the ciphersuites to SSL at init. This behavioral difference is
aligned in this PR.

### Testing:
A test to ensure changes to SSL_CTX don't impact SSL configurations and
we correctly inherit ciphersuites. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.