Migrate 3rd batch of CI jobs (#2183)
### Description of changes: 
This PR migrates jobs that require `ipv6` enabled in CodeBuild to run
properly (with `nmap` being a special add-on to balance out the size of
the next batch of jobs).

The jobs to be migrated in this PR are:
  - `grpc`
  - `bind9`
  - `strongswan`
  - `openvpn2.6`
  - `openvpnMaster`
  - `nmap`

### Call-outs:
In this PR, I changed the build directory of `bind9` to resolve a
character limit issue in CodeBuild.

In one of its tests, `bind9` will try to create a socket with a path
deep within the build directory. Since CodeBuild's build directory is
already very long, the combined socket path length will exceed the Linux
limit of 108 characters. Therefore, I had to reduce the length of the
build directory to mitigate this issue.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.