Add support to define a callback for FIPS test failures instead of aborting the process (#2162)
### Description of changes: 
This change adds a new build option `AWSLC_FIPS_FAILURE_CALLBACK` that
then requires applications that links with AWS-LC to define the
`AWS_LC_fips_failure_callback(const char* message)` function. If that
function is not defined AWS-LC fails to load.

When the callback is defined and a FIPS failure (CAST or PWCT) happens
AWS-LC calls that function instead. It is up to the calling application
to ensure no further cryptographic operations occur. This is not an
approved way to use AWS-LC in FIPS mode, for the approved way to build
and use AWS-LC see the latest FIPS security policy.

### Testing
This change adds a new gtest that checks the callback
function is called with the expected message and is used with
break-kat.go and the runtime PCT environment variables to ensure the
callback is passed the expected message.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.