Coverity Fix (#2236)
### Issues:
`P204971226`

### Description of changes: 
Add a explicit null check to ensure a valid leaf cert exists. We know in
current usage this will always be the case. Null check added for
future-proofing.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.