Enforce FIPS callback is only enabled for static builds (#2241)
### Description of changes: 
Due to how shared libraries and applications load it's not possible for
an application to dynamically link with AWS-LC and define it's own
custom callback before AWS-LC's constructor runs which expects the
callback to be defined. This change requires anyone expecting to build
with the callback to statically build and link AWS-LC so their callback
is defined before AWS-LC's constructor runs.

This change also enables the call_back_test to run all the time, this
test can run on any platform, platforms that don't support the callback
will get the default (abort) behavior.

This change also disables one death test if the callback is defined.
Because the callback doesn't call abort the test fails.

### Testing:
Update a test dimension to enable the callback and just run the tests
like normal to ensure everything passes.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.