Migrate last batch of jobs (#2214)
### Issues:
Resolves #CryptoAlg-2828

### Description of changes: 
Migrate the 4th and final batch of CI integration jobs:
- `python-main`
- `python-releases`
- `ruby-main`
- `ruby-releases`

Note: `libevent` is excluded due to some unknown errors with CodeBuild.
I will revisit it some time later.

### Call-outs:
1. CodeBuild does not allow combining `build-list` and `build-matrix`
(booo) so I have to manually list out the jobs that were supposed to be
in a matrix.
2. CodeBuild silences all `SIGINT` as `SIGIGN` so I added some patches
to override this behavior.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.