Make BIO_get_mem_data a function again (#2246)
### Description of changes: 
Can't track it down the historical reason why, but we have
`BIO_get_mem_data` defined as a preprocessor macro, but BoringSSL does
not. This creates an annoying compatibility issue for rust-openssl, and
has forced us to implement the function properly in our aws-lc-sys
crates in Rust. This redefines the function to have a concrete
definition.

As a side-note, the macro was technically wrong as it was casting
contents to a `char *` when it was actually `char **`, but this wasn't
problematic per-se since it was being passed into a `void *` argument.
If you look at the ctrl function definition you can see that it is
indeed treated as a `char **` pointer.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.