Add public wrapper to internal bn_minimal_width function (#2245)
### Description of changes: 
This PR adds a public wrapper for an existing internal function
bn_minimal_width. OpenSSL stores the minimal number of words needed in a
field called top within the bignum struct. AWS-LC does not do this,
instead we have a field called width which is a loose upper bound on
this value. The new function, BN_get_minimal_width will provide us with
the value of "top" for a given Bignum.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.