Add LibRdKafka to our CI (#2225)
### Issues:
`CryptoAlg-2863`

### Description of changes: 
Add librdkafka upstream to our CI. This was added as a Github action to allow running the librdkafka tests. These tests require spinning up a kafka cluster which was running into issues in docker/codebuild. 

The patchfile in this PR is only to allow running a specific testing target in librdkafka. We do not need this patchfile for compatibility. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.