Update EVP_PKEY ED keygen to use an internal function that can return the result of the PWCT (#2256)
### Description of changes: 
`ED25519_keypair` has always been void, this used to be okay because it
was infallible. Now it can fail due to the PWCT failing. Usually if that
happened the process would abort which is the default (and only
approved) behavior of `AWS_LC_FIPS_failure`. However, if AWS-LC is built
with the callback and that callback doesn't abort it is possible for
`EVP_PKEY_keygen` to continue. Changes:
* `EVP_PKEY_keygen` now detects this failure and return the error to the
caller
* `ED25519_keypair` doesn't increment the service indicator if the PWCT
fails

### Testing:
I updated the fips_callback_test to use the EVP API for ED and check the
return value

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.