Update patch for Postgres (#2232)
The original patch I had for Postgres wasn't the best, as it just simply
commented out the single test we didn't support. It's not a patch worthy
of submitting upstream.
This new patch simply adds a function check specific to AWS-LC and
leverages existing logic to skip the test. I've kept the error string
replacements out of the script to minimize churn with our Postgres CI
and left a comment regarding that.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.