Avoid double seed for non-UBE support runtimes (#2259)
In the non-UBE support runtime, a fresh lazy init will spawn a seed of the CTR-DRBG state. But then in rand_bytes_core(), there will be an immediate spawn of a reseed of the CTR-DRBG state to ensure uniqueness for non-UBE support.

Fix this by telling rand_bytes_core whether the CTR-DRBG state is fresh or not.