Minor build fixes for CMake and libssl on x86 (#2267)
### Description of changes: 
* Removes unnecessary variable evaluation in a handful of if statements
for `CMAKE_SYSTEM_PROCESSOR_LOWER`. These can result in a failure if
`CMAKE_SYSTEM_PROCESSOR_LOWER` is an empty string as the `${}` is
expanded first in the if-statement due to historical reasons.
* `libssl` can fail to compile if building on x86 since we don't pass
`-msse2` to the C++ compiler. This is due to the fact that the
`crypto/internal.h` is included by libssl and it explicitly checks for
SSE2 support. See
https://github.com/aws/aws-lc/blob/main/crypto/internal.h#L203-L214

### Testing
* Verified build with a modified `./util/32-bit-toolchain.cmake` that
only sets `-m32` (since building on an x86-64 platform).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.