Make AWS_LC_fips_failure_callback optional in builds with AWSLC_FIPS_FAILURE_CALLBACK (#2266)
### Description of changes: 
This makes the FIPS callback optional when AWS-LC is built with
AWSLC_FIPS_FAILURE_CALLBACK which turns it on. Previously if AWS-LC was
built with AWSLC_FIPS_FAILURE_CALLBACK then any application that might
link with AWS-LC would need to define the callback or AWS-LC would
abort. Turns out that's hard to ensure, this change goes back to the old
behavior:
* Regular FIPS build of AWS-LC doesn't require a callback and always
aborts
* FIPS build with AWSLC_FIPS_FAILURE_CALLBACK allows a callback to be
defined and AWS-LC will call that and not abort
* FIPS build with AWSLC_FIPS_FAILURE_CALLBACK and no callback defined
fallback to the first bullet point behavior


### Testing:
I created a new test target that partially replicates our internal FIPS
callback build.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.