ci: Nix flake and devShell (#2189)
### Issues:
none

### Description of changes: 

Adds a Nix package flake and a development shell with the tooling needed
to build and test aws-lc.

The flake allows other projects, such as
[s2n-tls](https://github.com/aws/s2n-tls), to directly pull aws-lc into
a nix Ci environment. The devshell is a self-contained development
shell, setup and managed by nix, focused on reproducible builds, and can
be helpful for bootstrapping new folks.

Example devshell usage:
```
% nix develop                                                                                                                                                                                                                       
Entering a devShell...              
[awslc nix] dougch@devdesktop22:~/gitrepos/aws-lc$ clean;configure;build;unit                                                
```

### Testing:

How is this change tested (unit tests, fuzz tests, etc.)? locally, new
action on my fork
Sample output of the GitHub Action:
https://github.com/dougch/aws-lc/actions/runs/13296817224/job/37130549429

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.