Apply additional X509 validation checks on certificates sourced from trust store  (#2230)
### Description of changes: 
Improves validation of certificates by also applying additional checks
for certificates retrieved from the trust store. This aligns AWS-LC with
OpenSSL 1.1.1 and later behavior.

### Testing:
Add a root certificate that does not have the expected basicConstraints
extension with the cA bit set to true. This certificate would be
rejected by OpenSSL, but is currently allowed to be trusted by AWS-LC
due to it being in the trust store.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.