Revert "Allow constructed strings in BER parsing (#2015)" (#2278)
This reverts commit 2a722260329a62e0682410060f84cf144960afe4.

### Issues:
Addresses `CryptoAlg-3037`

### Description of changes: 
We've ran into issues with parsing indefinite BER with PKCS7 and it
turns out our support for parsing is not as complete. Instead of parsing
BER to an invalid unusable state that's confusing, we should outright
disallow parsing of constructed strings in BER until we fix the issue.

### Call-outs:
N/A

### Testing:
Original tests.

Ruby CI is expected to fail. I'll update the patches in another PR.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.