Add the rehash utility to the openssl CLI tool (#2258)
### Description of changes: 
Adding the rehash utility to the OpenSSL CLI tool. This design makes a
number of different design choices from OpenSSL. Namely:
1) Multiple directory paths are not supported, therefore duplicate
filenames are not supported
2) Like OpenSSL, any existing symlinks matching the hash regex are
removed from the given directory. We do not provide an option to
preserve these.
3) We return a binary true or false to indicate status. OpenSSL counts
and returns the number of errors.
4) Simplified the general processing logic and hash table structures. 
5) Removed windows filesystem support, we only need this tool for unix
based systems.
6) We don't support the legacy hash format (MD5)

A wrapper bash script was created to call the openssl rehash
functionality. This script is made executable and placed in the install
folder via cmake. We also copy a test for this bash script to the build
dir.

### Testing:
Unit tests for the two tier hashtable and integration tests with symlink
generation. Added another test to verify the functionality of the
produced c_rehash script.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.