Update patches in Ruby CI (#2233)
We've finalized our changes upstream and we should update our patches to
previous branches to align. This also helps minimize churn when applying
these patches internally.

### Call-outs:
N/A

### Testing:
CI

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.