Reject DSA trailing garbage in EVP layer, add test cases (#2289)
# Description
The EVP API layer was not rejecting signatures that contained trailing
garbage after the `DSA_SIG` ASN.1 encoding (note it correctly reject
invalid signatures regardless). This did not match the behavior of the
`DSA_verify` and `DSA_check_signature` APIs which explicitly check and
reject trailing garbage. This aligns the EVP layer with this behavior.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.