Wire-up rust-openssl into GitHub CI (for the time being) (#2291)
### Description of changes: 
Adds validation that AWS-LC will build with rust-openssl when using the
bindgen feature or bindgen-cli if pointed at an AWS-LC installation
directory. We will add coverage on the aws-lc-rs side specifically for
handling coverage of the crate specific feature flag that toggles on
`aws-lc-sys`.

### Call-outs:
Not putting this into CodeBuild right now as it would require building
new docker images with Rust, and the current process is influx. Will
wait for the new deployment process to be completed so we can guarantee
more frequently updated rust compiler releases without having to
automate that with each invocation.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.