AES: Add function pointer trampoline to avoid delocator issue (#2294)
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

--------

On AArch64, the delocator can patch up the computation of function
pointers only if the pointers can be computed with a PC-relative offset
in the range (-1MB, 1MB).

For the function pointer computations in aes/mode_wrappers.c, this
bounds condition is about to be violated by further code additions to
AWS-LC, as witnessed in AES-unrelated PRs (specifically #2176).

This commit preventatively fixes the issue by adding function pointer
trampolines to crypto/fipsmodule/aes/mode_wrappers.c: These are stub
functions immediately branching into the desired assembly routines, but
close enough to the C code computing their address to ensure that their
addresses will be computable using a PC-relative offset.

This fix is similar to https://github.com/aws/aws-lc/pull/2165.

Mid/Long-term, it should be considered whether the delocator can
introduce the necessary indirections automatically.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>