Add more build options to match callback build (#2279)
### Resolves
CryptoAlg-2437

### Description of changes: 
Mirror more of the build options to make sure we catch any build issues
early in GitHub. Adding the release with debug issue raised an issue
with break-kat.go failing with: `Test input value was still found after
erasing it. Second copy?`. This happens because the function which
contains the test vectors are inlined into two functions which results
in two copies of the data. This only appears to happen with GCC 7.3.1.
This change fixes that by marking all of the tests as no inline.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.