Reorganizing compatability tests, rework certificates for better groking (#2305)
### Description of changes: 
This is mostly a re-organization PR so that the new comparability
related tests are in a separate unit file for easier grokking in the
future as more are added. I also re-generated the current certificates
added in the previous PR for the trust-store change to define a better
distinguished namespace hierarchy. Mainly to have better binding and
name-spacing of the root certificate to any intermediate, and end-entity
certificates generated for testing.

Additionally added a text representation of the certificates as well.
Note the certificate start dates are back date to 2015 through 2100.
This is mainly for consistency with the other certificates BoringSSL had
generated in various places in the library.

Otherwise this also adds a test case for RFC 5280 4.2.1.1 which states
with regards to the authority key identifier extension
> Conforming CAs MUST mark this extension as non-critical.

Previously this was unchecked on any certificates that were sourced from
the trust-store (so an intermediate for example could have this marked
as critical but would still be valid), and after making changes to
validating the [trust
store](https://github.com/aws/aws-lc/commit/1f367c97b007d4a85334aa6be90995eba67ad19b)
this test case in x509-limbo suddenly passed. So simply adding a similar
test case for ensuring that we actually validated this RFC statement.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.