Reinstate indefinite length and [UNIVERSAL 0] support in crypto/asn1 (#2306)
This is the first step of trying to reinstate support for BER in our
OpenSSL ASN1 macro parsers. Proper support for BER consists of two
parts, "indefinite length BER" and "implicitly tagged constructed BER
strings". This is being done to properly support BER in our PKCS7
parsers, as we've made the transition to switch from CBB -> ASN1 macros
for PKCS7 for now: 621bceb. We were using `CBS_asn1_ber_to_der` to
translate BER -> DER for `d2i_PKCS7`, but the function has a caveat with
implicitly-tagged constructed strings that causes encoding translations to
be incomplete. We've decided to restore BER support in the macro parsers
to resolve the PKCS7 compatibility issue and provide better OpenSSL
compatibility for possible applications using OpenSSL's ASN.1 functions
with BER encoding.

This commit is a revert of the two prior commits below. All "new logic"
in `crypto/asn1/tasn_dec.c` was taken from these two commits.
* edbdc24
* ee510f5

The only smaller new changes are the following:
1. I've only translated the comment code styling and added a few more
comments clarifying some of the historic ASN1 shenanigans.
2. There's also a change to the call to
`asn1_get_object_maybe_indefinite` to call `ASN1_get_object`. This
aligns with the original code in OpenSSL. I think we can clean up some
of the changes done in cfe86a3, but I'd prefer to do so in a separate
commit.

### Testing:
* I've reused the BER tests in our CBB parsers here so we're confident
that the functionalities 1:1. This commit is reinstating indefinite
length so I've only borrowed those corresponding tests for now.
* I've generated a few new ASN1 test files with indefinite length as
well. Indefinite length is more common among container types like SET
and SEQUENCE, so we only pass those tests for now. Other indefinite
length scenarios involve constructed form, but we don't have support for
that just yet.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.