Implemented no-op CRYPTO_mem_ctrl (#2295)
### Issues:
Addresses #CryptoAlg-2905

### Description of changes: 
Applications requiring CRYPTO_mem_ctrl can now build successfully.
However, we define `OPENSSL_NO_CRYPTO_MDEBUG` in AWS-LC, which disables
this functionality. As a result, the implemented function intentionally
does nothing.

### Call-outs:
We've added a defensive compilation error that will trigger if
`OPENSSL_NO_CRYPTO_MDEBUG` is removed in the future. This will allow us
to fully implement the function if needed.

### Testing:
No testing is required for this intentional no-operation function. This
approach is consistent with OpenSSL 1.1.1's handling of
`OPENSSL_NO_CRYPTO_MDEBUG`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: kingstjo <kingstjo@users.noreply.github.com>