Reinstate support for constructed strings in crypto/asn1 (#2310)
This is the last step of trying to reinstate support for BER in our
OpenSSL ASN1 macro parsers. Proper support for BER consists of two
parts, "indefinite length BER" and "implicitly tagged constructed BER
strings". The first step of supporting indefinite length BER can be
found in eafd940.

This commit is a revert of the prior commit below. All "new logic" in
`crypto/asn1/tasn_dec.c` was taken from this.
* a70edd4

The only smaller new changes are the following:
* The original commit uses an additional `free_cont` parameter to pass
into `asn1_ex_c2i`. This indicates whether a buffer should be reused or
newly allocated. This made things a bit funky and the actual performance
upgrade for reusing a buffer seems negligible. Removing the use of
`free_cont` altogether helps isolate the ASN1 code changes to just one
static function in `crypto/asn1/tasn_dec.c` and cleans up logic.

### Testing:
* I've reused the BER tests in our CBB parsers here so we're confident
that the functionalities 1:1.
eafd940 only borrowed the ones that were indefinite length, while this
commit takes and tests against constructed string test cases.
* eafd940 documented some test cases that were failing due to no
support for constructed strings. I've update the documentation and
turned on the tests, as this commit now allows proper support of them.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.