Additional X.509 Behavior Compatability Tests (#2312)
* CRL Distribution Points extension per RFC 5280 should be marked
non-critical, but OpenSSL 1.1.1 does allow for it to be marked critical
by a CA issuing a certificate. Thus we modify to allow for this
condition and add an associated test-case.
* Adds a test for a Root CA missing the basic constraints extension and
a corresponding end-entity certificate signed by such a root.
* Adds a test for a certificate containing a negative serial number.
While this does not conform to RFC 5280, it is supported by OpenSSL
1.1.1 and many other implementations. This adds a test-case for this
behavior which was not previously covered.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.