Fix bug with GCC 11.5 and check_test_optional_abort, add CI for AL2023 with GCC 11.5 (#2311)
### Issues
Fixes P218664081

### Description of changes: 
Cover a small testing gap for an AL2023 customers that want to use the
callback feature.

### Callouts
Why change where a stack variable is defined? This works around a
compiler bug with GCC 11.5 that was causing the incorrect result from
memcmp to be returned and the integrity check for some algorithms to get
the wrong result. P218664081

This is a different GCC version with a subtly different version of the
memcmp bug from https://github.com/aws/aws-lc/pull/73. In this case we
can fix our code to work around the issue and we have tests that ensure
all the integrity and CAST tests work as expected.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.