Add null check for EVP_get_digestbyobj (#2360)
Some usages of Ruby were failing internally with a NULL pointer passed
to this function. Ruby OpenSSL's usage tries to parse a string and
obtain an `EVP_MD` from it, but a NULL pointer is passed when an invalid
string is used.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.