Randomness generation KATs (#2334)
Implements KAT tests for randomness generation. That is, we override the entropy vtable, compute randomness, and then verify the generated randomness is equal to what is expected.

This proves that we can predict the generated randomness. In turn, proving that a specific entropy method is used where expected and the output is used as input into the CTR-DRBG functions as expected.