Revert "Update patch for Postgres (#2232)" (#2358)
This reverts commit db723116a144386007295521987feed4a6daab2f.

The original patch was much more stable. Instead of changing the
postgres configure script to check against AWS-LC functions, we comment
out the relevant test instead. I've added comments in the patch to
reference back to the commit if we ever decide to upstream support to
postgres.
The Postgres ARM CI dimension is failing, but the failure doesn't seem
relevant to this patch.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.